hide the country code from URL
