+++ src/app/(main)/categories/[...category]/page.tsx
