+++ src/app/(main)/results/[query]/page.tsx
